requirements.txt
pandas>=2.1
matplotlib>=3.8
pyyaml
requests
pytz
python-dateutil
kubernetes